+++ integrationtest/mapper/orm/src/test/java/org/hibernate/search/integrationtest/mapper/orm/automaticindexing/association/bytype/manytomany/ownedbycontaining/AutomaticIndexingManyToManyOwnedByContainingListBaseIT.java
@@ -4,30 +4,26 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.multi;
+package org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.bytype.manytomany.ownedbycontaining;
 
 import java.util.ArrayList;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Stream;
 import javax.persistence.Basic;
 import javax.persistence.ElementCollection;
 import javax.persistence.Entity;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
 import javax.persistence.JoinTable;
 import javax.persistence.ManyToMany;
-import javax.persistence.MapKeyColumn;
 import javax.persistence.OneToOne;
 import javax.persistence.OrderBy;
 import javax.persistence.Transient;
 
-import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.AbstractAutomaticIndexingAssociationIT;
-import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.ContainerPrimitives;
-import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.MultiValuedPropertyAccessor;
-import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.SingleValuedPropertyAccessor;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.bytype.AbstractAutomaticIndexingMultiValuedAssociationBaseIT;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.bytype.ContainerPrimitives;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.bytype.MultiValuedPropertyAccessor;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.bytype.SingleValuedPropertyAccessor;
 import org.hibernate.search.mapper.pojo.automaticindexing.ReindexOnUpdate;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.GenericField;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
@@ -37,29 +33,26 @@ import org.hibernate.search.mapper.pojo.mapping.definition.annotation.ObjectPath
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.PropertyValue;
 
 /**
- * Test automatic indexing based on Hibernate ORM entity events
- * when a Map-values association is involved.
- * <p>
- * See {@link AbstractAutomaticIndexingAssociationIT} for more details on how this test is designed.
+ * Test automatic indexing caused by multi-valued association updates
+ * or by updates of associated (contained) entities,
+ * with a {@code @ManyToMany List} association owned by the containing side.
  */
-public class AutomaticIndexingMapValuesAssociationIT extends AbstractAutomaticIndexingMultiAssociationIT<
-		AutomaticIndexingMapValuesAssociationIT.IndexedEntity,
-		AutomaticIndexingMapValuesAssociationIT.ContainingEntity,
-		AutomaticIndexingMapValuesAssociationIT.ContainedEntity,
-		Map<String, AutomaticIndexingMapValuesAssociationIT.ContainedEntity>,
-		List<AutomaticIndexingMapValuesAssociationIT.ContainingEntity>
-		> {
+public class AutomaticIndexingManyToManyOwnedByContainingListBaseIT
+		extends AbstractAutomaticIndexingMultiValuedAssociationBaseIT<
+						AutomaticIndexingManyToManyOwnedByContainingListBaseIT.IndexedEntity,
+						AutomaticIndexingManyToManyOwnedByContainingListBaseIT.ContainingEntity,
+						AutomaticIndexingManyToManyOwnedByContainingListBaseIT.ContainedEntity,
+						List<AutomaticIndexingManyToManyOwnedByContainingListBaseIT.ContainedEntity>,
+						List<AutomaticIndexingManyToManyOwnedByContainingListBaseIT.ContainingEntity>
+				> {
 
-	public AutomaticIndexingMapValuesAssociationIT() {
-		super( new ModelPrimitives() );
+	public AutomaticIndexingManyToManyOwnedByContainingListBaseIT() {
+		super( new ModelPrimitivesImpl() );
 	}
 
-	private static class ModelPrimitives
-			implements MultiAssociationModelPrimitives<IndexedEntity, ContainingEntity, ContainedEntity,
-						Map<String, ContainedEntity>, List<ContainingEntity>> {
-
-		private final ContainerPrimitives<Map<String, ContainedEntity>, ContainedEntity> MAP_VALUES_PRIMITIVES =
-				ContainerPrimitives.mapValues( containedEntity -> containedEntity.getIndexedField() + "_value" );
+	private static class ModelPrimitivesImpl
+			implements MultiValuedModelPrimitives<IndexedEntity, ContainingEntity, ContainedEntity,
+					List<ContainedEntity>, List<ContainingEntity>> {
 
 		@Override
 		public String getIndexName() {
@@ -119,14 +112,13 @@ public class AutomaticIndexingMapValuesAssociationIT extends AbstractAutomaticIn
 		}
 
 		@Override
-		public Map<String, ContainedEntity> newContainedAssociation(Map<String, ContainedEntity> original) {
-			return new LinkedHashMap<>( original );
+		public List<ContainedEntity> newContainedAssociation(List<ContainedEntity> original) {
+			return new ArrayList<>( original );
 		}
 
 		@Override
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<String, ContainedEntity>> containedIndexedEmbedded() {
-			return new MultiValuedPropertyAccessor<>(
-					MAP_VALUES_PRIMITIVES,
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, List<ContainedEntity>> containedIndexedEmbedded() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
 					ContainingEntity::getContainedIndexedEmbedded, ContainingEntity::setContainedIndexedEmbedded );
 		}
 
@@ -137,9 +129,8 @@ public class AutomaticIndexingMapValuesAssociationIT extends AbstractAutomaticIn
 		}
 
 		@Override
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<String, ContainedEntity>> containedNonIndexedEmbedded() {
-			return new MultiValuedPropertyAccessor<>(
-					MAP_VALUES_PRIMITIVES,
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, List<ContainedEntity>> containedNonIndexedEmbedded() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
 					ContainingEntity::getContainedNonIndexedEmbedded, ContainingEntity::setContainedNonIndexedEmbedded );
 		}
 
@@ -150,26 +141,24 @@ public class AutomaticIndexingMapValuesAssociationIT extends AbstractAutomaticIn
 		}
 
 		@Override
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<String, ContainedEntity>> containedIndexedEmbeddedShallowReindexOnUpdate() {
-			return new MultiValuedPropertyAccessor<>(
-					MAP_VALUES_PRIMITIVES,
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, List<ContainedEntity>> containedIndexedEmbeddedShallowReindexOnUpdate() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
 					ContainingEntity::getContainedIndexedEmbeddedShallowReindexOnUpdate,
 					ContainingEntity::setContainedIndexedEmbeddedShallowReindexOnUpdate );
 		}
 
 		@Override
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<String, ContainedEntity>> containedIndexedEmbeddedNoReindexOnUpdate() {
-			return new MultiValuedPropertyAccessor<>(
-					MAP_VALUES_PRIMITIVES,
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, List<ContainedEntity>> containedIndexedEmbeddedNoReindexOnUpdate() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
 					ContainingEntity::getContainedIndexedEmbeddedNoReindexOnUpdate,
 					ContainingEntity::setContainedIndexedEmbeddedNoReindexOnUpdate );
 		}
 
 		@Override
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<String, ContainedEntity>> containedUsedInCrossEntityDerivedProperty() {
-			return new MultiValuedPropertyAccessor<>(
-					MAP_VALUES_PRIMITIVES,
-					ContainingEntity::getContainedUsedInCrossEntityDerivedProperty );
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, List<ContainedEntity>> containedUsedInCrossEntityDerivedProperty() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
+					ContainingEntity::getContainedUsedInCrossEntityDerivedProperty,
+					ContainingEntity::setContainedUsedInCrossEntityDerivedProperty );
 		}
 
 		@Override
@@ -180,10 +169,9 @@ public class AutomaticIndexingMapValuesAssociationIT extends AbstractAutomaticIn
 
 		@Override
 		@SuppressWarnings("unchecked")
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<String, ContainedEntity>> containedIndexedEmbeddedWithCast() {
-			return new MultiValuedPropertyAccessor<>(
-					MAP_VALUES_PRIMITIVES,
-					root -> (Map) root.getContainedIndexedEmbeddedWithCast() );
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, List<ContainedEntity>> containedIndexedEmbeddedWithCast() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
+					root -> (List) root.getContainedIndexedEmbeddedWithCast() );
 		}
 
 		@Override
@@ -266,70 +254,34 @@ public class AutomaticIndexingMapValuesAssociationIT extends AbstractAutomaticIn
 		private ContainingEntity child;
 
 		@ManyToMany
-		@JoinTable(
-				name = "indexed_containedIndexedEmbedded",
-				joinColumns = @JoinColumn(name = "mapHolder"),
-				inverseJoinColumns = @JoinColumn(name = "value")
-		)
-		@MapKeyColumn(name = "map_key")
+		@JoinTable(name = "indexed_containedIndexedEmbedded")
 		@IndexedEmbedded(includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" })
-		@OrderBy("id asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		private Map<String, ContainedEntity> containedIndexedEmbedded = new LinkedHashMap<>();
+		private List<ContainedEntity> containedIndexedEmbedded = new ArrayList<>();
 
 		@ManyToMany
-		@JoinTable(
-				name = "indexed_containedNonIndexedEmbedded",
-				joinColumns = @JoinColumn(name = "mapHolder"),
-				inverseJoinColumns = @JoinColumn(name = "value")
-		)
-		@MapKeyColumn(name = "map_key")
-		@OrderBy("id asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		private Map<String, ContainedEntity> containedNonIndexedEmbedded = new LinkedHashMap<>();
+		@JoinTable(name = "indexed_containedNonIndexedEmbedded")
+		private List<ContainedEntity> containedNonIndexedEmbedded = new ArrayList<>();
 
 		@ManyToMany
-		@JoinTable(
-				name = "indexed_containedIndexedEmbeddedShallowReindexOnUpdate",
-				joinColumns = @JoinColumn(name = "mapHolder"),
-				inverseJoinColumns = @JoinColumn(name = "value")
-		)
-		@MapKeyColumn(name = "map_key")
+		@JoinTable(name = "indexed_indexedEmbeddedShallowReindexOnUpdateContained")
 		@IndexedEmbedded(includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" })
 		@IndexingDependency(reindexOnUpdate = ReindexOnUpdate.SHALLOW)
-		@OrderBy("id asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		private Map<String, ContainedEntity> containedIndexedEmbeddedShallowReindexOnUpdate = new LinkedHashMap<>();
+		private List<ContainedEntity> containedIndexedEmbeddedShallowReindexOnUpdate = new ArrayList<>();
 
 		@ManyToMany
-		@JoinTable(
-				name = "indexed_containedIndexedEmbeddedNoReindexOnUpdate",
-				joinColumns = @JoinColumn(name = "mapHolder"),
-				inverseJoinColumns = @JoinColumn(name = "value")
-		)
-		@MapKeyColumn(name = "map_key")
+		@JoinTable(name = "indexed_indexedEmbeddedNoReindexOnUpdateContained")
 		@IndexedEmbedded(includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" })
 		@IndexingDependency(reindexOnUpdate = ReindexOnUpdate.NO)
-		@OrderBy("id asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		private Map<String, ContainedEntity> containedIndexedEmbeddedNoReindexOnUpdate = new LinkedHashMap<>();
+		private List<ContainedEntity> containedIndexedEmbeddedNoReindexOnUpdate = new ArrayList<>();
 
 		@ManyToMany
-		@JoinTable(
-				name = "indexed_containedUsedInCrossEntityDerivedProperty",
-				joinColumns = @JoinColumn(name = "mapHolder"),
-				inverseJoinColumns = @JoinColumn(name = "value")
-		)
-		@MapKeyColumn(name = "map_key")
-		@OrderBy("id asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		private Map<String, ContainedEntity> containedUsedInCrossEntityDerivedProperty = new LinkedHashMap<>();
+		@JoinTable(name = "indexed_containedUsedInCrossEntityDerivedProperty")
+		private List<ContainedEntity> containedUsedInCrossEntityDerivedProperty = new ArrayList<>();
 
 		@ManyToMany(targetEntity = ContainedEntity.class)
-		@JoinTable(
-				name = "indexed_containedIndexedEmbeddedWithCast",
-				joinColumns = @JoinColumn(name = "mapHolder"),
-				inverseJoinColumns = @JoinColumn(name = "value")
-		)
-		@MapKeyColumn(name = "map_key")
+		@JoinTable(name = "indexed_containedIndexedEmbeddedWithCast")
 		@IndexedEmbedded(includePaths = "indexedField", targetType = ContainedEntity.class)
-		@OrderBy("id asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		private Map<String, Object> containedIndexedEmbeddedWithCast = new LinkedHashMap<>();
+		private List<Object> containedIndexedEmbeddedWithCast = new ArrayList<>();
 
 		public Integer getId() {
 			return id;
@@ -363,45 +315,50 @@ public class AutomaticIndexingMapValuesAssociationIT extends AbstractAutomaticIn
 			this.child = child;
 		}
 
-		public Map<String, ContainedEntity> getContainedIndexedEmbedded() {
+		public List<ContainedEntity> getContainedIndexedEmbedded() {
 			return containedIndexedEmbedded;
 		}
 
-		public void setContainedIndexedEmbedded(Map<String, ContainedEntity> containedIndexedEmbedded) {
+		public void setContainedIndexedEmbedded(List<ContainedEntity> containedIndexedEmbedded) {
 			this.containedIndexedEmbedded = containedIndexedEmbedded;
 		}
 
-		public Map<String, ContainedEntity> getContainedNonIndexedEmbedded() {
+		public List<ContainedEntity> getContainedNonIndexedEmbedded() {
 			return containedNonIndexedEmbedded;
 		}
 
-		public void setContainedNonIndexedEmbedded(Map<String, ContainedEntity> containedNonIndexedEmbedded) {
+		public void setContainedNonIndexedEmbedded(List<ContainedEntity> containedNonIndexedEmbedded) {
 			this.containedNonIndexedEmbedded = containedNonIndexedEmbedded;
 		}
 
-		public Map<String, ContainedEntity> getContainedIndexedEmbeddedShallowReindexOnUpdate() {
+		public List<ContainedEntity> getContainedIndexedEmbeddedShallowReindexOnUpdate() {
 			return containedIndexedEmbeddedShallowReindexOnUpdate;
 		}
 
 		public void setContainedIndexedEmbeddedShallowReindexOnUpdate(
-				Map<String, ContainedEntity> containedIndexedEmbeddedShallowReindexOnUpdate) {
+				List<ContainedEntity> containedIndexedEmbeddedShallowReindexOnUpdate) {
 			this.containedIndexedEmbeddedShallowReindexOnUpdate = containedIndexedEmbeddedShallowReindexOnUpdate;
 		}
 
-		public Map<String, ContainedEntity> getContainedIndexedEmbeddedNoReindexOnUpdate() {
+		public List<ContainedEntity> getContainedIndexedEmbeddedNoReindexOnUpdate() {
 			return containedIndexedEmbeddedNoReindexOnUpdate;
 		}
 
 		public void setContainedIndexedEmbeddedNoReindexOnUpdate(
-				Map<String, ContainedEntity> containedIndexedEmbeddedNoReindexOnUpdate) {
+				List<ContainedEntity> containedIndexedEmbeddedNoReindexOnUpdate) {
 			this.containedIndexedEmbeddedNoReindexOnUpdate = containedIndexedEmbeddedNoReindexOnUpdate;
 		}
 
-		public Map<String, ContainedEntity> getContainedUsedInCrossEntityDerivedProperty() {
+		public List<ContainedEntity> getContainedUsedInCrossEntityDerivedProperty() {
 			return containedUsedInCrossEntityDerivedProperty;
 		}
 
-		public Map<String, Object> getContainedIndexedEmbeddedWithCast() {
+		public void setContainedUsedInCrossEntityDerivedProperty(
+				List<ContainedEntity> containedUsedInCrossEntityDerivedProperty) {
+			this.containedUsedInCrossEntityDerivedProperty = containedUsedInCrossEntityDerivedProperty;
+		}
+
+		public List<Object> getContainedIndexedEmbeddedWithCast() {
 			return containedIndexedEmbeddedWithCast;
 		}
 
@@ -419,7 +376,7 @@ public class AutomaticIndexingMapValuesAssociationIT extends AbstractAutomaticIn
 		})
 		public Optional<String> getCrossEntityDerivedField() {
 			return computeDerived(
-					containedUsedInCrossEntityDerivedProperty.values().stream().flatMap( c -> Stream.of(
+					containedUsedInCrossEntityDerivedProperty.stream().flatMap( c -> Stream.of(
 							c.getFieldUsedInCrossEntityDerivedField1(),
 							c.getFieldUsedInCrossEntityDerivedField2()
 					) )
@@ -585,5 +542,4 @@ public class AutomaticIndexingMapValuesAssociationIT extends AbstractAutomaticIn
 			return computeDerived( Stream.of( fieldUsedInContainedDerivedField1, fieldUsedInContainedDerivedField2 ) );
 		}
 	}
-
 }
