+++ integrationtest/mapper/orm/src/test/java/org/hibernate/search/integrationtest/mapper/orm/automaticindexing/association/bytype/manytomany/ownedbycontaining/AutomaticIndexingManyToManyOwnedByContainingSortedSetBaseIT.java
@@ -4,67 +4,60 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.multi;
+package org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.bytype.manytomany.ownedbycontaining;
 
 import java.util.ArrayList;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
+import java.util.SortedSet;
+import java.util.TreeSet;
 import java.util.stream.Stream;
 import javax.persistence.Basic;
-import javax.persistence.Column;
 import javax.persistence.ElementCollection;
 import javax.persistence.Entity;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
 import javax.persistence.JoinTable;
 import javax.persistence.ManyToMany;
-import javax.persistence.MapKeyClass;
-import javax.persistence.MapKeyJoinColumn;
 import javax.persistence.OneToOne;
 import javax.persistence.OrderBy;
 import javax.persistence.Transient;
 
-import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.AbstractAutomaticIndexingAssociationIT;
-import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.ContainerPrimitives;
-import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.MultiValuedPropertyAccessor;
-import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.SingleValuedPropertyAccessor;
+import org.hibernate.annotations.SortNatural;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.bytype.AbstractAutomaticIndexingMultiValuedAssociationBaseIT;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.bytype.ContainerPrimitives;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.bytype.MultiValuedPropertyAccessor;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.bytype.SingleValuedPropertyAccessor;
 import org.hibernate.search.mapper.pojo.automaticindexing.ReindexOnUpdate;
-import org.hibernate.search.mapper.pojo.extractor.builtin.BuiltinContainerExtractors;
-import org.hibernate.search.mapper.pojo.mapping.definition.annotation.AssociationInverseSide;
-import org.hibernate.search.mapper.pojo.extractor.mapping.annotation.ContainerExtraction;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.GenericField;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.IndexedEmbedded;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.IndexingDependency;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.ObjectPath;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.PropertyValue;
+import org.hibernate.search.util.impl.test.annotation.TestForIssue;
 
 /**
- * Test automatic indexing based on Hibernate ORM entity events
- * when a Map-keys association is involved.
- * <p>
- * See {@link AbstractAutomaticIndexingAssociationIT} for more details on how this test is designed.
+ * Test automatic indexing caused by multi-valued association updates
+ * or by updates of associated (contained) entities,
+ * with a {@code @ManyToMany SortedSet} association owned by the containing side.
  */
-public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticIndexingMultiAssociationIT<
-		AutomaticIndexingMapKeysAssociationIT.IndexedEntity,
-		AutomaticIndexingMapKeysAssociationIT.ContainingEntity,
-		AutomaticIndexingMapKeysAssociationIT.ContainedEntity,
-		Map<AutomaticIndexingMapKeysAssociationIT.ContainedEntity, String>,
-		List<AutomaticIndexingMapKeysAssociationIT.ContainingEntity>
-		> {
+@TestForIssue(jiraKey = "HSEARCH-2490")
+public class AutomaticIndexingManyToManyOwnedByContainingSortedSetBaseIT
+		extends AbstractAutomaticIndexingMultiValuedAssociationBaseIT<
+						AutomaticIndexingManyToManyOwnedByContainingSortedSetBaseIT.IndexedEntity,
+						AutomaticIndexingManyToManyOwnedByContainingSortedSetBaseIT.ContainingEntity,
+						AutomaticIndexingManyToManyOwnedByContainingSortedSetBaseIT.ContainedEntity,
+						SortedSet<AutomaticIndexingManyToManyOwnedByContainingSortedSetBaseIT.ContainedEntity>,
+						List<AutomaticIndexingManyToManyOwnedByContainingSortedSetBaseIT.ContainingEntity>
+				> {
 
-	public AutomaticIndexingMapKeysAssociationIT() {
-		super( new ModelPrimitives() );
+	public AutomaticIndexingManyToManyOwnedByContainingSortedSetBaseIT() {
+		super( new ModelPrimitivesImpl() );
 	}
 
-	private static class ModelPrimitives
-			implements MultiAssociationModelPrimitives<IndexedEntity, ContainingEntity, ContainedEntity,
-						Map<ContainedEntity, String>, List<ContainingEntity>> {
-
-		private final ContainerPrimitives<Map<ContainedEntity, String>, ContainedEntity> MAP_KEYS_PRIMITIVES =
-				ContainerPrimitives.mapKeys( containedEntity -> containedEntity.getIndexedField() + "_value" );
+	private static class ModelPrimitivesImpl
+			implements MultiValuedModelPrimitives<IndexedEntity, ContainingEntity, ContainedEntity,
+											SortedSet<ContainedEntity>, List<ContainingEntity>> {
 
 		@Override
 		public String getIndexName() {
@@ -124,13 +117,13 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		}
 
 		@Override
-		public Map<ContainedEntity, String> newContainedAssociation(Map<ContainedEntity, String> original) {
-			return new LinkedHashMap<>( original );
+		public SortedSet<ContainedEntity> newContainedAssociation(SortedSet<ContainedEntity> original) {
+			return new TreeSet<>( original );
 		}
 
 		@Override
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<ContainedEntity, String>> containedIndexedEmbedded() {
-			return new MultiValuedPropertyAccessor<>( MAP_KEYS_PRIMITIVES,
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, SortedSet<ContainedEntity>> containedIndexedEmbedded() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
 					ContainingEntity::getContainedIndexedEmbedded, ContainingEntity::setContainedIndexedEmbedded );
 		}
 
@@ -141,8 +134,8 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		}
 
 		@Override
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<ContainedEntity, String>> containedNonIndexedEmbedded() {
-			return new MultiValuedPropertyAccessor<>( MAP_KEYS_PRIMITIVES,
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, SortedSet<ContainedEntity>> containedNonIndexedEmbedded() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
 					ContainingEntity::getContainedNonIndexedEmbedded, ContainingEntity::setContainedNonIndexedEmbedded );
 		}
 
@@ -153,22 +146,22 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		}
 
 		@Override
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<ContainedEntity, String>> containedIndexedEmbeddedShallowReindexOnUpdate() {
-			return new MultiValuedPropertyAccessor<>( MAP_KEYS_PRIMITIVES,
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, SortedSet<ContainedEntity>> containedIndexedEmbeddedShallowReindexOnUpdate() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
 					ContainingEntity::getContainedIndexedEmbeddedShallowReindexOnUpdate,
 					ContainingEntity::setContainedIndexedEmbeddedShallowReindexOnUpdate );
 		}
 
 		@Override
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<ContainedEntity, String>> containedIndexedEmbeddedNoReindexOnUpdate() {
-			return new MultiValuedPropertyAccessor<>( MAP_KEYS_PRIMITIVES,
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, SortedSet<ContainedEntity>> containedIndexedEmbeddedNoReindexOnUpdate() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
 					ContainingEntity::getContainedIndexedEmbeddedNoReindexOnUpdate,
 					ContainingEntity::setContainedIndexedEmbeddedNoReindexOnUpdate );
 		}
 
 		@Override
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<ContainedEntity, String>> containedUsedInCrossEntityDerivedProperty() {
-			return new MultiValuedPropertyAccessor<>( MAP_KEYS_PRIMITIVES,
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, SortedSet<ContainedEntity>> containedUsedInCrossEntityDerivedProperty() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
 					ContainingEntity::getContainedUsedInCrossEntityDerivedProperty );
 		}
 
@@ -180,9 +173,9 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 
 		@Override
 		@SuppressWarnings("unchecked")
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<ContainedEntity, String>> containedIndexedEmbeddedWithCast() {
-			return new MultiValuedPropertyAccessor<>( MAP_KEYS_PRIMITIVES,
-					root -> (Map) root.getContainedIndexedEmbeddedWithCast() );
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, SortedSet<ContainedEntity>> containedIndexedEmbeddedWithCast() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
+					root -> (SortedSet) root.getContainedIndexedEmbeddedWithCast() );
 		}
 
 		@Override
@@ -264,91 +257,41 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		})
 		private ContainingEntity child;
 
-		@ElementCollection
-		@JoinTable(
-				name = "indexed_containedIndexedEmbedded",
-				joinColumns = @JoinColumn(name = "mapHolder")
-		)
-		@MapKeyJoinColumn(name = "map_key")
-		@Column(name = "value")
-		@OrderBy("map_key asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		@IndexedEmbedded(
-				includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" },
-				extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-		)
-		private Map<ContainedEntity, String> containedIndexedEmbedded = new LinkedHashMap<>();
+		@ManyToMany
+		@JoinTable(name = "indexed_containedIndexedEmbedded")
+		@IndexedEmbedded(includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" })
+		@SortNatural
+		private SortedSet<ContainedEntity> containedIndexedEmbedded = new TreeSet<>();
 
-		@ElementCollection
-		@JoinTable(
-				name = "indexed_containedNonIndexedEmbedded",
-				joinColumns = @JoinColumn(name = "mapHolder")
-		)
-		@MapKeyJoinColumn(name = "map_key")
-		@Column(name = "value")
-		@OrderBy("map_key asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		private Map<ContainedEntity, String> containedNonIndexedEmbedded = new LinkedHashMap<>();
+		@ManyToMany
+		@JoinTable(name = "indexed_containedNonIndexedEmbedded")
+		@SortNatural
+		private SortedSet<ContainedEntity> containedNonIndexedEmbedded = new TreeSet<>();
 
-		@ElementCollection
-		@JoinTable(
-				name = "indexed_containedIndexedEmbeddedShallowReindexOnUpdate",
-				joinColumns = @JoinColumn(name = "mapHolder")
-		)
-		@MapKeyJoinColumn(name = "map_key")
-		@Column(name = "value")
-		@OrderBy("map_key asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		@IndexedEmbedded(
-				includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" },
-				extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-		)
-		@IndexingDependency(
-				reindexOnUpdate = ReindexOnUpdate.SHALLOW,
-				extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-		)
-		private Map<ContainedEntity, String> containedIndexedEmbeddedShallowReindexOnUpdate = new LinkedHashMap<>();
+		@ManyToMany
+		@JoinTable(name = "indexed_indexedEmbeddedShallowReindexOnUpdateContained")
+		@IndexedEmbedded(includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" })
+		@IndexingDependency(reindexOnUpdate = ReindexOnUpdate.SHALLOW)
+		@SortNatural
+		private SortedSet<ContainedEntity> containedIndexedEmbeddedShallowReindexOnUpdate = new TreeSet<>();
 
-		@ElementCollection
-		@JoinTable(
-				name = "indexed_containedIndexedEmbeddedNoReindexOnUpdate",
-				joinColumns = @JoinColumn(name = "mapHolder")
-		)
-		@MapKeyJoinColumn(name = "map_key")
-		@Column(name = "value")
-		@OrderBy("map_key asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		@IndexedEmbedded(
-				includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" },
-				extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-		)
-		@IndexingDependency(
-				reindexOnUpdate = ReindexOnUpdate.NO,
-				extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-		)
-		private Map<ContainedEntity, String> containedIndexedEmbeddedNoReindexOnUpdate = new LinkedHashMap<>();
+		@ManyToMany
+		@JoinTable(name = "indexed_indexedEmbeddedNoReindexOnUpdateContained")
+		@IndexedEmbedded(includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" })
+		@IndexingDependency(reindexOnUpdate = ReindexOnUpdate.NO)
+		@SortNatural
+		private SortedSet<ContainedEntity> containedIndexedEmbeddedNoReindexOnUpdate = new TreeSet<>();
 
-		@ElementCollection
-		@JoinTable(
-				name = "indexed_containedUsedInCrossEntityDerivedProperty",
-				joinColumns = @JoinColumn(name = "mapHolder")
-		)
-		@MapKeyJoinColumn(name = "map_key")
-		@Column(name = "value")
-		@OrderBy("map_key asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		private Map<ContainedEntity, String> containedUsedInCrossEntityDerivedProperty = new LinkedHashMap<>();
+		@ManyToMany
+		@JoinTable(name = "indexed_containedUsedInCrossEntityDerivedProperty")
+		@SortNatural
+		private SortedSet<ContainedEntity> containedUsedInCrossEntityDerivedProperty = new TreeSet<>();
 
-		@ElementCollection
-		@JoinTable(
-				name = "indexed_containedIndexedEmbeddedWithCast",
-				joinColumns = @JoinColumn(name = "mapHolder")
-		)
-		@MapKeyClass(ContainedEntity.class)
-		@MapKeyJoinColumn(name = "map_key")
-		@Column(name = "value")
-		@OrderBy("map_key asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		@IndexedEmbedded(
-				includePaths = { "indexedField" },
-				extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY),
-				targetType = ContainedEntity.class
-		)
-		private Map<Object, String> containedIndexedEmbeddedWithCast = new LinkedHashMap<>();
+		@ManyToMany(targetEntity = ContainedEntity.class)
+		@JoinTable(name = "indexed_containedIndexedEmbeddedWithCast")
+		@IndexedEmbedded(includePaths = "indexedField", targetType = ContainedEntity.class)
+		@SortNatural
+		private SortedSet<Object> containedIndexedEmbeddedWithCast = new TreeSet<>();
 
 		public Integer getId() {
 			return id;
@@ -382,45 +325,45 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 			this.child = child;
 		}
 
-		public Map<ContainedEntity, String> getContainedIndexedEmbedded() {
+		public SortedSet<ContainedEntity> getContainedIndexedEmbedded() {
 			return containedIndexedEmbedded;
 		}
 
-		public void setContainedIndexedEmbedded(Map<ContainedEntity, String> containedIndexedEmbedded) {
+		public void setContainedIndexedEmbedded(SortedSet<ContainedEntity> containedIndexedEmbedded) {
 			this.containedIndexedEmbedded = containedIndexedEmbedded;
 		}
 
-		public Map<ContainedEntity, String> getContainedNonIndexedEmbedded() {
+		public SortedSet<ContainedEntity> getContainedNonIndexedEmbedded() {
 			return containedNonIndexedEmbedded;
 		}
 
-		public void setContainedNonIndexedEmbedded(Map<ContainedEntity, String> containedNonIndexedEmbedded) {
+		public void setContainedNonIndexedEmbedded(SortedSet<ContainedEntity> containedNonIndexedEmbedded) {
 			this.containedNonIndexedEmbedded = containedNonIndexedEmbedded;
 		}
 
-		public Map<ContainedEntity, String> getContainedIndexedEmbeddedShallowReindexOnUpdate() {
+		public SortedSet<ContainedEntity> getContainedIndexedEmbeddedShallowReindexOnUpdate() {
 			return containedIndexedEmbeddedShallowReindexOnUpdate;
 		}
 
 		public void setContainedIndexedEmbeddedShallowReindexOnUpdate(
-				Map<ContainedEntity, String> containedIndexedEmbeddedShallowReindexOnUpdate) {
+				SortedSet<ContainedEntity> containedIndexedEmbeddedShallowReindexOnUpdate) {
 			this.containedIndexedEmbeddedShallowReindexOnUpdate = containedIndexedEmbeddedShallowReindexOnUpdate;
 		}
 
-		public Map<ContainedEntity, String> getContainedIndexedEmbeddedNoReindexOnUpdate() {
+		public SortedSet<ContainedEntity> getContainedIndexedEmbeddedNoReindexOnUpdate() {
 			return containedIndexedEmbeddedNoReindexOnUpdate;
 		}
 
 		public void setContainedIndexedEmbeddedNoReindexOnUpdate(
-				Map<ContainedEntity, String> containedIndexedEmbeddedNoReindexOnUpdate) {
+				SortedSet<ContainedEntity> containedIndexedEmbeddedNoReindexOnUpdate) {
 			this.containedIndexedEmbeddedNoReindexOnUpdate = containedIndexedEmbeddedNoReindexOnUpdate;
 		}
 
-		public Map<ContainedEntity, String> getContainedUsedInCrossEntityDerivedProperty() {
+		public SortedSet<ContainedEntity> getContainedUsedInCrossEntityDerivedProperty() {
 			return containedUsedInCrossEntityDerivedProperty;
 		}
 
-		public Map<Object, String> getContainedIndexedEmbeddedWithCast() {
+		public SortedSet<Object> getContainedIndexedEmbeddedWithCast() {
 			return containedIndexedEmbeddedWithCast;
 		}
 
@@ -428,23 +371,17 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		@GenericField
 		@IndexingDependency(derivedFrom = {
 				@ObjectPath({
-						@PropertyValue(
-								propertyName = "containedUsedInCrossEntityDerivedProperty",
-								extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-						),
+						@PropertyValue(propertyName = "containedUsedInCrossEntityDerivedProperty"),
 						@PropertyValue(propertyName = "fieldUsedInCrossEntityDerivedField1")
 				}),
 				@ObjectPath({
-						@PropertyValue(
-								propertyName = "containedUsedInCrossEntityDerivedProperty",
-								extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-						),
+						@PropertyValue(propertyName = "containedUsedInCrossEntityDerivedProperty"),
 						@PropertyValue(propertyName = "fieldUsedInCrossEntityDerivedField2")
 				})
 		})
 		public Optional<String> getCrossEntityDerivedField() {
-			return AbstractAutomaticIndexingAssociationIT.computeDerived(
-					containedUsedInCrossEntityDerivedProperty.keySet().stream().flatMap( c -> Stream.of(
+			return computeDerived(
+					containedUsedInCrossEntityDerivedProperty.stream().flatMap( c -> Stream.of(
 							c.getFieldUsedInCrossEntityDerivedField1(),
 							c.getFieldUsedInCrossEntityDerivedField2()
 					) )
@@ -461,68 +398,25 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 	}
 
 	@Entity(name = "contained")
-	public static class ContainedEntity {
+	public static class ContainedEntity implements Comparable<ContainedEntity> {
 
 		@Id
 		private Integer id;
 
-		/*
-		 * No mappedBy here. The inverse side of associations modeled by a Map key cannot use mappedBy.
-		 * If they do, Hibernate assumes that map *values* are the opposite side of the association,
-		 * and ends up adding all kind of wrong foreign keys.
-		 */
-		@ManyToMany
-		@JoinTable(name = "contained_mapHolder")
+		@ManyToMany(mappedBy = "containedIndexedEmbedded")
 		@OrderBy("id asc") // Make sure the iteration order is predictable
-		@AssociationInverseSide(
-				inversePath = @ObjectPath(
-						@PropertyValue(
-								propertyName = "containedIndexedEmbedded",
-								extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-						)
-				)
-		)
 		private List<ContainingEntity> containingAsIndexedEmbedded = new ArrayList<>();
 
-		/*
-		 * No mappedBy here, same reasons as above.
-		 */
-		@ManyToMany
-		@JoinTable(name = "contained_nonIndexedMapHolder")
+		@ManyToMany(mappedBy = "containedNonIndexedEmbedded")
 		@OrderBy("id asc") // Make sure the iteration order is predictable
 		private List<ContainingEntity> containingAsNonIndexedEmbedded = new ArrayList<>();
 
-		/*
-		 * No mappedBy here, same reasons as above.
-		 */
-		@ManyToMany
+		@ManyToMany(mappedBy = "containedUsedInCrossEntityDerivedProperty")
 		@OrderBy("id asc") // Make sure the iteration order is predictable
-		@AssociationInverseSide(
-				inversePath = @ObjectPath(
-						@PropertyValue(
-								propertyName = "containedUsedInCrossEntityDerivedProperty",
-								extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-						)
-				)
-		)
 		private List<ContainingEntity> containingAsUsedInCrossEntityDerivedProperty = new ArrayList<>();
 
-		/*
-		 * No mappedBy here. The inverse side of associations modeled by a Map key cannot use mappedBy.
-		 * If they do, Hibernate assumes that map *values* are the opposite side of the association,
-		 * and ends up adding all kind of wrong foreign keys.
-		 */
-		@ManyToMany(targetEntity = ContainingEntity.class)
-		@JoinTable(name = "contained_withCastMapHolder")
+		@ManyToMany(mappedBy = "containedIndexedEmbeddedWithCast", targetEntity = ContainingEntity.class)
 		@OrderBy("id asc") // Make sure the iteration order is predictable
-		@AssociationInverseSide(
-				inversePath = @ObjectPath(
-						@PropertyValue(
-								propertyName = "containedIndexedEmbeddedWithCast",
-								extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-						)
-				)
-		)
 		private List<Object> containingAsIndexedEmbeddedWithCast = new ArrayList<>();
 
 		@Basic
@@ -554,6 +448,11 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 
 		@Basic // Do not annotate with @GenericField, this would make the test pointless
 		private String fieldUsedInCrossEntityDerivedField2;
+
+		@Override
+		public int compareTo(ContainedEntity o) {
+			return getId() - o.getId();
+		}
 
 		public Integer getId() {
 			return id;
@@ -650,8 +549,7 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 				@ObjectPath(@PropertyValue(propertyName = "fieldUsedInContainedDerivedField2"))
 		})
 		public Optional<String> getContainedDerivedField() {
-			return AbstractAutomaticIndexingAssociationIT.computeDerived( Stream.of( fieldUsedInContainedDerivedField1, fieldUsedInContainedDerivedField2 ) );
+			return computeDerived( Stream.of( fieldUsedInContainedDerivedField1, fieldUsedInContainedDerivedField2 ) );
 		}
 	}
-
 }
